Added SQL file to create database views and used it in logs analysis
__Problem__
It is taking quite long time to fetch the answers to the first two questions.

__Fix__
According to a suggestion provided during code review it would be a better idea to store the list of paths, with a count of
the number of times they occur in the log table in a database `VIEW` instead of trying to perform comparision on the log table which is
relatively massive (1,677,735 rows!).